requirements.txt
ollama
pandas
Pillow
glob2